Generate payment batch / delete batch
Summing is jan up to the given month. Is that right? Actually create
the batch, payments, and corresponding cause balance records, then
produce a CSV with all info needed for payment. Add check_num to
payment; for now it equals the payment id. Allow deleting a batch,
which deletes and payments and adjustments, and adjusts the cause
balance records.