Jobs GUI for importing transactions
New DelayedRake model stores arguments and a job_identifier (set to the
Delayed::Job. For each type of job, check if there are any running. If
none are running, show the form for starting one. If one is running,
hide the form and show a link to the progress page. This page shows
running and failed jobs, with arguments and status info.